@@ -1,14 +1,14 @@
 package com.cpacm.musicbtn;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import com.cpacm.FloatingMusicButton;
 import com.cpacm.FloatingMusicMenu;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class FmmActivity extends AppCompatActivity {
 
@@ -21,15 +21,15 @@ public class FmmActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_fmm);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
         extraFab = new FloatingMusicButton(this);
         extraFab.setCoverDrawable(getResources().getDrawable(R.drawable.author));
 
-        upFmm = (FloatingMusicMenu) findViewById(R.id.fmm);
-        addFab = (FloatingActionButton) findViewById(R.id.add_fab);
-        subFab = (FloatingActionButton) findViewById(R.id.sub_fab);
+        upFmm = findViewById(R.id.fmm);
+        addFab = findViewById(R.id.add_fab);
+        subFab = findViewById(R.id.sub_fab);
         addFab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
