@@ -6,15 +6,18 @@ import android.media.AudioManager;
 import android.media.MediaPlayer;
 import android.os.Handler;
 import android.os.Message;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
+import android.widget.ImageView;
 import android.widget.SeekBar;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.cpacm.FloatingMusicMenu;
+import com.cpacm.RotatingProgressDrawable;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.io.IOException;
 
@@ -28,6 +31,8 @@ public class MusicPlayerActivity extends AppCompatActivity implements MediaPlaye
     private MusicHandler musicHandler;
     private boolean isPlaying = false;
 
+    private ImageView rotIv;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -35,12 +40,12 @@ public class MusicPlayerActivity extends AppCompatActivity implements MediaPlaye
 
         musicHandler = new MusicHandler();
 
-        progressTv = (TextView) findViewById(R.id.progress_tv);
-        totalTv = (TextView) findViewById(R.id.total_tv);
-        musicSeekBar = (SeekBar) findViewById(R.id.music_seekbar);
-        fmm = (FloatingMusicMenu) findViewById(R.id.fmm);
-        playOrPauseFab = (FloatingActionButton) findViewById(R.id.fab_play);
-        nextFab = (FloatingActionButton) findViewById(R.id.fab_next);
+        progressTv = findViewById(R.id.progress_tv);
+        totalTv = findViewById(R.id.total_tv);
+        musicSeekBar = findViewById(R.id.music_seekbar);
+        fmm = findViewById(R.id.fmm);
+        playOrPauseFab = findViewById(R.id.fab_play);
+        nextFab = findViewById(R.id.fab_next);
         nextFab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
